# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Accounting & Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties:
    get:
      operationId: listParties
      tags: [parties]
      summary: List all parties
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [customer, supplier, both]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Party"
    post:
      operationId: createParty
      tags: [parties]
      summary: Create party
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartyBody"
      responses:
        "201":
          description: Party created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{id}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get party by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
        "404":
          description: Not found
    patch:
      operationId: updateParty
      tags: [parties]
      summary: Update party
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartyBody"
      responses:
        "200":
          description: Party updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    delete:
      operationId: deleteParty
      tags: [parties]
      summary: Delete party
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /parties/{id}/ledger:
    get:
      operationId: getPartyLedger
      tags: [parties]
      summary: Party ledger transactions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Party ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyLedger"

  # LEDGERS
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order / status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/convert:
    post:
      operationId: convertOrderToInvoice
      tags: [orders]
      summary: Convert order to sale invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleInvoice"

  # PURCHASE INVOICES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Party:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
        openingBalance:
          type: number
        balanceType:
          type: string
          enum: [dr, cr]
        currentBalance:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - openingBalance
        - balanceType
        - createdAt
    CreatePartyBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
        openingBalance:
          type: number
        balanceType:
          type: string
          enum: [dr, cr]
      required:
        - name
        - type
        - openingBalance
        - balanceType
    UpdatePartyBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
        openingBalance:
          type: number
        balanceType:
          type: string
          enum: [dr, cr]
    PartyLedger:
      type: object
      properties:
        party:
          $ref: "#/components/schemas/Party"
        openingBalance:
          type: number
        openingType:
          type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/LedgerTransaction"
        closingBalance:
          type: number
        closingType:
          type: string
      required:
        - party
        - openingBalance
        - openingType
        - transactions
        - closingBalance
        - closingType
    LedgerTransaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
        description:
          type: string
        debit:
          type: ["number", "null"]
        credit:
          type: ["number", "null"]
        balance:
          type: number
        balanceType:
          type: string
      required:
        - id
        - date
        - type
        - description
        - balance
        - balanceType
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        total:
          type: number
      required:
        - id
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
        - taxableAmount
        - cgst
        - sgst
        - igst
        - total
    InvoiceItemInput:
      type: object
      properties:
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
      required:
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
    PaymentMode:
      type: object
      properties:
        mode:
          type: string
          enum: [cash, upi, cheque, bank_transfer, credit]
        amount:
          type: number
        reference:
          type: ["string", "null"]
      required:
        - mode
        - amount
    SaleInvoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        subtotal:
          type: number
        totalDiscount:
          type: number
        totalTaxable:
          type: number
        totalCgst:
          type: number
        totalSgst:
          type: number
        totalIgst:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
        amountPaid:
          type: number
        balanceDue:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, confirmed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - date
        - partyName
        - isGst
        - isInterstate
        - subtotal
        - totalDiscount
        - totalTaxable
        - totalCgst
        - totalSgst
        - totalIgst
        - totalGst
        - grandTotal
        - amountPaid
        - balanceDue
        - status
        - items
        - payments
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyPhone:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, dispatched, delivered, cancelled]
        grandTotal:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        convertedInvoiceId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - date
        - partyName
        - status
        - grandTotal
        - items
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyPhone:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required:
        - date
        - partyName
        - items
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, dispatched, delivered, cancelled]
        notes:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
...